+++ categories.json
@@ -5,6 +5,7 @@
       {
         "name": "Filter",
         "classes": [
+          "Backgrid.Extension.AbstractFilter",
           "Backgrid.Extension.ServerSideFilter",
           "Backgrid.Extension.ClientSideFilter",
           "Backgrid.Extension.LunrFilter"
